Extract functions to init_jinja_globals  module
