+++ airflow/www/app.py
@@ -16,11 +16,9 @@
 # specific language governing permissions and limitations
 # under the License.
 #
-import socket
 from datetime import timedelta
 from typing import Optional
 
-import pendulum
 from flask import Flask
 from flask_appbuilder import SQLA
 from flask_caching import Cache
@@ -31,6 +29,7 @@ from airflow.configuration import conf
 from airflow.logging_config import configure_logging
 from airflow.utils.json import AirflowJsonEncoder
 from airflow.www.extensions.init_appbuilder import init_appbuilder
+from airflow.www.extensions.init_jinja_globals import init_jinja_globals
 from airflow.www.extensions.init_session import init_logout_timeout, init_permanent_session
 from airflow.www.extensions.init_views import (
     init_api_connexion, init_api_experimental, init_appbuilder_views, init_error_handlers, init_flash_views,
@@ -119,46 +118,7 @@ def create_app(config=None, testing=False, app_name="Airflow"):
     configure_manifest_files(app)
 
     with app.app_context():
-        def init_jinja_globals(app):
-            server_timezone = conf.get('core', 'default_timezone')
-            if server_timezone == "system":
-                server_timezone = pendulum.local_timezone().name
-            elif server_timezone == "utc":
-                server_timezone = "UTC"
 
-            default_ui_timezone = conf.get('webserver', 'default_ui_timezone')
-            if default_ui_timezone == "system":
-                default_ui_timezone = pendulum.local_timezone().name
-            elif default_ui_timezone == "utc":
-                default_ui_timezone = "UTC"
-            if not default_ui_timezone:
-                default_ui_timezone = server_timezone
-
-            @app.context_processor
-            def jinja_globals():  # pylint: disable=unused-variable
-
-                globals = {
-                    'server_timezone': server_timezone,
-                    'default_ui_timezone': default_ui_timezone,
-                    'hostname': socket.getfqdn() if conf.getboolean(
-                        'webserver', 'EXPOSE_HOSTNAME', fallback=True) else 'redact',
-                    'navbar_color': conf.get(
-                        'webserver', 'NAVBAR_COLOR'),
-                    'log_fetch_delay_sec': conf.getint(
-                        'webserver', 'log_fetch_delay_sec', fallback=2),
-                    'log_auto_tailing_offset': conf.getint(
-                        'webserver', 'log_auto_tailing_offset', fallback=30),
-                    'log_animation_speed': conf.getint(
-                        'webserver', 'log_animation_speed', fallback=1000)
-                }
-
-                if 'analytics_tool' in conf.getsection('webserver'):
-                    globals.update({
-                        'analytics_tool': conf.get('webserver', 'ANALYTICS_TOOL'),
-                        'analytics_id': conf.get('webserver', 'ANALYTICS_ID')
-                    })
-
-                return globals
 
         def init_xframe_protection(app):
             @app.after_request
